+++ src/plugin/Pipe_Plugin.h
@@ -1,0 +1,10 @@
+#ifndef SHOW__PIPE_PLUGIN_H
+#define SHOW__PIPE_PLUGIN_H
+
+
+#include "Plugin.h"
+
+extern Plugin Pipe_Plugin;
+
+
+#endif